+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `@Platform(exclude=...)` annotation value to remove header files from inherited `@Platform(include=...`
  * Fix a few issues with `Parser`, including missing `PointerPointer` member setters ([issue bytedeco/javacpp-presets#478](https://github.com/bytedeco/javacpp-presets/issues/478))
  * Fix potential race conditions and various issues with `Loader` that could prevent libraries like MKL from working properly
  * Add `Loader.addressof()` to access native symbols, usable via optional `ValueGetter/ValueSetter` in `FunctionPointer`